Finally did it.
Implemented fall through mechanic as well as crumbling platform mechanic.

Fall through is attached to individual objects, so not all objects will be fall though.  Can use different animated block to signify this.

Crumbling platform is weird, and needs work to fix the visual discrepency.  Needs another type of material to signify that it is "collapsed," then might need some editing so that it doesn't return whenever the object's "dimention" is not the one the player is currently switched too.